+++ bridge_adapters/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "bridge_adapters"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+slvm = { path = "../vm" }
+compile_state = { path = "../compile_state" }
+bridge_types = { workspace = true }
