+++ builtins/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 slvm = { path = "../vm" }
 compile_state = { path = "../compile_state" }
+bridge_adapters = { path = "../bridge_adapters" }
 unicode-segmentation = "1.10.1"
 bridge_types = { workspace = true }
 bridge_macros = { path = "../bridge_macros" }
