+++ builtins/src/collections.rs
@@ -1,5 +1,5 @@
 use crate::SloshVm;
-use compile_state::add_builtin;
+use bridge_adapters::add_builtin;
 use slvm::{VMError, VMResult, Value};
 
 pub fn vec_slice(vm: &mut SloshVm, registers: &[Value]) -> VMResult<Value> {
