+++ builtins/src/io.rs
@@ -1,4 +1,4 @@
-use compile_state::add_builtin;
+use bridge_adapters::add_builtin;
 use compile_state::state::SloshVm;
 use slvm::{VMError, VMResult, Value};
 use std::collections::HashMap;
