+++ slosh/Cargo.toml
@@ -10,6 +10,7 @@ lisp-test = ["dep:regex", "dep:lazy_static"]
 [dependencies]
 unicode_reader = "1"
 sl-compiler = { path = "../compiler" }
+bridge_adapters = { path = "../bridge_adapters" }
 sl-liner = { git = "https://github.com/sl-sh-dev/sl-liner.git" }
 slvm = { path = "../vm" }
 builtins = { path = "../builtins" }
